pyproject.toml
# PyTorch is now automatically selected based on platform:
# - CPU builds on macOS/Windows
# - CUDA builds on Linux
# - Nightly CUDA builds on Linux (aarch64)
#
# For manual control, you can still use: UV_TORCH_BACKEND=auto uv sync

[project]
name = "sae-lens"
version = "5.9.1"
description = "Training and Analyzing Sparse Autoencoders (SAEs)"
authors = [
    { name = "Joseph Bloom" }
]
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = [
    "deep-learning",
    "sparse-autoencoders",
    "mechanistic-interpretability",
    "PyTorch",
]
classifiers = [
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "transformer-lens>=2.0.0",
    "transformers>=4.38.1",
    "plotly>=5.19.0",
    "plotly-express>=0.4.1",
    "matplotlib>=3.8.3",
    "matplotlib-inline>=0.1.6",
    "datasets>=2.17.1",
    "babe>=0.0.7",
    "nltk>=3.8.1",
    "safetensors>=0.4.2",
    "typer>=0.12.3",
    "pyzmq==26.0.0",
    "automated-interpretability>=0.0.5,<1.0.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "pytest-profiling>=1.7.0",
    "zstandard>=0.22.0",
    "typing-extensions>=4.10.0",
    "simple-parsing>=0.1.6",
    # New dependencies from hypersteer
    "einops>=0.8.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "seaborn>=0.13.0",
    "tiktoken>=0.5.0",
    "wandb>=0.16.0",
    "together>=1.0.0",
    "hydra-colorlog>=1.0.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.0.0",
    "tqdm>=4.65.0",
    "httpx>=0.24.0",
    "openai>=1.0.0",
    "pyvene>=0.1.0",
    "adjusttext>=1.0.0",
    "optuna==4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.2",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.6.2",
    "pyright==1.1.365",
    "ansible-lint>=24.2.3; platform_system != 'Windows'",
    "botocore>=1.34.101",
    "boto3>=1.34.101",
    "docstr-coverage>=2.3.2",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.5.34",
    "mkdocs-autorefs>=1.1.0",
    "mkdocs-section-index>=0.3.9",
    "mkdocstrings>=0.25.2",
    "mkdocstrings-python>=1.10.9",
    "tabulate>=0.9.0",
    "ruff>=0.7.4",
    "eai-sparsify>=1.1.1",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "ai_commit @ git+https://github.com/sidnb13/toolbox.git#subdirectory=src/ai_commit",
    "hydra-ray-jobs-launcher @ git+https://github.com/sidnb13/hydra#egg=hydra-ray-jobs-launcher&subdirectory=plugins/hydra_ray_jobs_launcher",
]
mamba = [
    "mamba-lens>=0.0.4",
]
flash-attn = [
    "flash-attn>=2.0.0",
]

[project.urls]
Homepage = "https://jbloomaus.github.io/SAELens"
Repository = "https://github.com/jbloomAus/SAELens"
Documentation = "https://jbloomaus.github.io/SAELens"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["sae_lens"]
artifacts = ["pretrained_saes.yaml"]

[tool.ruff.lint]
exclude = ["*.ipynb"]
ignore = ["E203", "E501", "E731", "F722", "E741", "F821", "F403", "ARG002"]
select = ["UP", "TID", "I", "F", "E", "ARG", "SIM", "RET", "LOG", "T20"]

[tool.ruff.lint.per-file-ignores]
"benchmark/*" = ["T20"]
"scripts/*" = ["T20"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"typing.Union".msg = "Use `|` instead"
"typing.Optional".msg = "Use `| None` instead"
"typing.Dict".msg = "Use `dict` instead"
"typing.Tuple".msg = "Use `tuple` instead"
"typing.List".msg = "Use `list` instead"

[tool.pyright]
typeCheckingMode = "strict"
reportMissingTypeStubs = "none"
reportUnknownMemberType = "none"
reportUnknownArgumentType = "none"
reportUnknownVariableType = "none"
reportUntypedFunctionDecorator = "none"
reportUnnecessaryIsInstance = "none"
reportUnnecessaryComparison = "none"
reportConstantRedefinition = "none"
reportUnknownLambdaType = "none"
reportPrivateUsage = "none"
reportDeprecated = "none"
reportPrivateImportUsage = "none"
ignore = ["**/wandb/**"]
